# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Orders System API",
        "description": "API для системы обработки заказов",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/v1/api-docs": {
            "get": {
                "tags": [
                    "Documentation"
                ],
                "summary": "Получение документации API",
                "operationId": "3494d5d46c5de99541b2dbc2480adb0b",
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "openapi": {
                                            "type": "string",
                                            "example": "3.0.0"
                                        },
                                        "info": {
                                            "type": "object"
                                        },
                                        "paths": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Получение списка заказов",
                "operationId": "5894f7ae19272fefaf73e3a6f9d9782d",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Количество записей на странице",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список заказов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/OrderResponse"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Создание нового заказа",
                "operationId": "fccc497d8bcac727afdc854cf0521698",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Заказ успешно создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Сообщение об ошибке",
                                            "type": "string"
                                        },
                                        "errors": {
                                            "description": "Список ошибок валидации",
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Получение информации о заказе",
                "operationId": "37b9054f8bb7c660ae6ee13bfb030a08",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID заказа",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о заказе",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Заказ не найден"
                    }
                }
            }
        },
        "/orders/{id}/status": {
            "patch": {
                "tags": [
                    "Orders"
                ],
                "summary": "Обновление статуса заказа",
                "operationId": "1b652569febe1298d5c611a716763852",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID заказа",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "processing",
                                            "completed",
                                            "cancelled"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Статус заказа обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Заказ не найден"
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Сообщение об ошибке",
                                            "type": "string"
                                        },
                                        "errors": {
                                            "description": "Список ошибок валидации",
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "OrderRequest": {
                "required": [
                    "customer_name",
                    "customer_email",
                    "items"
                ],
                "properties": {
                    "customer_name": {
                        "description": "Имя клиента",
                        "type": "string"
                    },
                    "customer_email": {
                        "description": "Email клиента",
                        "type": "string",
                        "format": "email"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "required": [
                                "product_name",
                                "quantity",
                                "price"
                            ],
                            "properties": {
                                "product_name": {
                                    "description": "Название товара",
                                    "type": "string"
                                },
                                "quantity": {
                                    "description": "Количество",
                                    "type": "integer",
                                    "minimum": 1
                                },
                                "price": {
                                    "description": "Цена за единицу",
                                    "type": "number",
                                    "format": "float",
                                    "minimum": 0
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "OrderResponse": {
                "properties": {
                    "data": {
                        "properties": {
                            "id": {
                                "description": "ID заказа",
                                "type": "integer"
                            },
                            "customer_name": {
                                "description": "Имя клиента",
                                "type": "string"
                            },
                            "customer_email": {
                                "description": "Email клиента",
                                "type": "string",
                                "format": "email"
                            },
                            "status": {
                                "description": "Статус заказа",
                                "type": "string",
                                "enum": [
                                    "pending",
                                    "processing",
                                    "completed",
                                    "cancelled"
                                ]
                            },
                            "total_amount": {
                                "description": "Общая сумма заказа",
                                "type": "number",
                                "format": "float"
                            },
                            "items": {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "id": {
                                            "description": "ID позиции заказа",
                                            "type": "integer"
                                        },
                                        "product_name": {
                                            "description": "Название товара",
                                            "type": "string"
                                        },
                                        "quantity": {
                                            "description": "Количество",
                                            "type": "integer"
                                        },
                                        "price": {
                                            "description": "Цена за единицу",
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "total": {
                                            "description": "Общая стоимость позиции",
                                            "type": "number",
                                            "format": "float"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "created_at": {
                                "description": "Дата создания",
                                "type": "string",
                                "format": "date-time"
                            },
                            "updated_at": {
                                "description": "Дата обновления",
                                "type": "string",
                                "format": "date-time"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Documentation",
            "description": "Documentation"
        },
        {
            "name": "Orders",
            "description": "Orders"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ]
}